+++ Panther/include/sema/conversion.hpp
@@ -27,7 +27,7 @@ namespace pcit::panther{
 namespace pcit::panther::sema{
 
 
-	[[nodiscard]] auto exprToGenericValue(Expr expr, const class panther::Context& context) -> core::GenericValue;
+	[[nodiscard]] auto exprToGenericValue(Expr expr, class panther::Context& context) -> core::GenericValue;
 
 	[[nodiscard]] auto extractStringIDFromExpr(Expr expr, const class panther::Context& context) -> StringValueID;
 	[[nodiscard]] auto extractStringFromExpr(Expr expr, const class panther::Context& context) -> std::string_view;